refactor: reorganize generated code structure and update imports
- Move generated GraphQL code from src/@generated to src/generated
- Update Prisma generator output paths and client imports
- Change docs generator output to top-level docs directory
- Remove @generated TypeScript path mappings
- Update all import statements to use new generated paths
- Add docs directory to gitignore
- Install @types/passport-jwt for TypeScript support
