+++ package.json
@@ -69,6 +69,7 @@
     "@types/jest": "^30.0.0",
     "@types/multer": "^2.0.0",
     "@types/node": "^24.7.0",
+    "@types/passport-jwt": "^4.0.1",
     "@types/supertest": "^6.0.3",
     "eslint": "^9.37.0",
     "eslint-config-prettier": "^10.1.8",
