+++ src/board-followers/board-followers.service.ts
@@ -1,5 +1,5 @@
 import { PrismaService } from '@/prisma/prisma.service';
-import { BoardFollower, Prisma } from '@generated/prisma';
+import { BoardFollower, Prisma } from 'generated/prisma';
 import { Injectable, NotFoundException } from '@nestjs/common';
 
 @Injectable()
