+++ src/categories/categories.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { Prisma, Category } from '@generated/prisma';
+import { Prisma, Category } from 'generated/prisma';
 import { PrismaService } from '@/prisma/prisma.service';
 
 @Injectable()
