+++ src/types/graphql/index.ts
@@ -1,5 +1,5 @@
 import { Field, ID } from '@nestjs/graphql';
-import { User } from 'generated/graphql';
+import { User } from '@/generated/graphql';
 
 export * from './login.input';
 export * from './auth.payload';
