+++ src/users/users.resolver.ts
@@ -5,10 +5,10 @@ import {
   FindUniqueUserArgs,
   UpdateOneUserArgs,
   User,
-} from '@/@generated/graphql';
+} from '@/generated/graphql';
 import { JwtAuthGuard } from '@/utils/guards';
 import { UseGuards } from '@nestjs/common';
-import { Prisma } from '@generated/prisma';
+import { Prisma } from 'generated/prisma';
 import { type GraphQLResolveInfo } from 'graphql';
 import { PrismaSelect } from '@paljs/plugins';
