+++ tsconfig.json
@@ -16,8 +16,6 @@
     "baseUrl": "./",
     "paths": {
       "@/*": ["src/*"],
-      "@generated/*": ["src/@generated/*"],
-      "@generated/prisma": ["generated/prisma"]
     },
     "incremental": true,
     "skipLibCheck": true,
